refactor: demo folders structure
