+++ src/screensets/drafts/demo/components/DataDisplayElements.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Avatar, AvatarImage, AvatarFallback, Badge } from '@hai3/uikit';
 import { useTranslation, TextLoader } from '@hai3/uicore';
-import { BadgeCheckIcon } from '../../../uikit/icons/BadgeCheckIcon';
-import { DEMO_SCREENSET_ID } from '../../../demoScreenset';
-import { UI_KIT_ELEMENTS_SCREEN_ID } from '../UIKitElementsScreen';
+import { BadgeCheckIcon } from '../uikit/icons/BadgeCheckIcon';
+import { DEMO_SCREENSET_ID } from '../demoScreenset';
+import { UI_KIT_ELEMENTS_SCREEN_ID } from '../screens/uikit/UIKitElementsScreen';
 
 /**
  * Data Display Elements Component
